// src/app/_dto/tbldocattch.dto.ts
export class TblDocAttachDto {

    HDocS_SysID: number = 0;
    HDoc_Tab_ID: number = 0;
    HDoc_SysID: number = 0;
    HDoc_GridNo: number = null;

    HDoc_Code: string = null;
    HDoc_Name: string = null;
    HDoc_Path: string = null;
    HDoc_Start_Date: Date = null;
    HDoc_Stop_Date: Date = null;
    HDoc_Expiry_Date: Date = null;
    HDoc_Alert_Date: Date = null;


    HDoc_Emp_SysID: number = 0;
    HDoc_Emp_Code: string = null;
    HDoc_Emp_Name: string = null;
    HDoc_Emp_Email: string = null;


    HDoc_Alert_SysID: number = 0;
    HDoc_Alert_Code: string = null;
    HDoc_Alert_Name: string = null;
    HDoc_Alert_Stop_Hrs: string = null;
    HDoc_Alert_Stop_Yn: string = null;
    HDoc_AlertRep_Hrs: string = null;
}